autorization de formation et de entreprise

+++ core/app/Http/Controllers/EntrepriseController.php
@@ -6,14 +6,17 @@ use App\Models\Entreprise;
 use App\Http\Requests\StoreEntrepriseRequest;
 use App\Http\Requests\UpdateEntrepriseRequest;
 use App\Http\Resources\EntrepriseResource;
+use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
 
 class EntrepriseController extends Controller
 {
+    use AuthorizesRequests;
     /**
      * Display a listing of the resource.
      */
     public function index()
     {
+        $this->authorize('viewAny');
         return EntrepriseResource::collection(Entreprise::all());
     }
 
@@ -22,7 +25,8 @@ class EntrepriseController extends Controller
      */
     public function store(StoreEntrepriseRequest $request)
     {
-        $formFields = $request ->validated();
+        $this->authorize('create');
+        $formFields = $request->validated();
         $entreprise = Entreprise::create($formFields);
         $response = new EntrepriseResource($entreprise);
         return response()->json([
@@ -36,6 +40,7 @@ class EntrepriseController extends Controller
      */
     public function show(Entreprise $entreprise)
     {
+        $this->authorize('view', $entreprise);
         return new EntrepriseResource($entreprise);   
     }
 
@@ -44,7 +49,7 @@ class EntrepriseController extends Controller
      */
     public function update(UpdateEntrepriseRequest $request, Entreprise $entreprise)
     {
-        
+        $this->authorize('update', $entreprise);
         $formFields =$request->validated();
         $entreprise->update($formFields);
         return response()->json([
@@ -58,6 +63,7 @@ class EntrepriseController extends Controller
      */
     public function destroy(Entreprise $entreprise)
     {
+        $this->authorize('delete', $entreprise);
         $entreprise->delete();
         return response()->json([
             'entreprise' => $entreprise,
