+++ core/app/Http/Controllers/FormationController.php
@@ -6,14 +6,17 @@ use App\Models\Formation;
 use App\Http\Requests\StoreFormationRequest;
 use App\Http\Requests\UpdateFormationRequest;
 use App\Http\Resources\FormationResource;
+use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
 
 class FormationController extends Controller
 {
+    use AuthorizesRequests;
     /**
      * Display a listing of the resource.
      */
     public function index()
     {
+        $this->authorize('viewAny');
         return FormationResource::collection(Formation::all());
     }
 
@@ -22,6 +25,7 @@ class FormationController extends Controller
      */
     public function store(StoreFormationRequest $request)
     {
+        $this->authorize('create');
         $formFields = $request ->validated();
         $formation = Formation::create($formFields);
         $response = new FormationResource($formation);
@@ -36,6 +40,7 @@ class FormationController extends Controller
      */
     public function show(Formation $formation)
     {
+        $this->authorize('view', $formation);
         return new FormationResource($formation); 
     }
 
@@ -44,6 +49,7 @@ class FormationController extends Controller
      */
     public function update(UpdateFormationRequest $request, Formation $formation)
     {
+        $this->authorize('update', $formation);
         $formation->update($request->validated());
         return response()->json([
             'entreprise' => new FormationResource($formation),
@@ -56,6 +62,7 @@ class FormationController extends Controller
      */
     public function destroy(Formation $formation)
     {
+        $this->authorize('delete', $formation);
         $formation->delete();
         return response()->json([
             'entreprise' => $formation,
